% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvQuantile.R
\name{cvQuantile}
\alias{cvQuantile}
\title{cvQuantile}
\usage{
cvQuantile(B, pv, m, alpha)
}
\arguments{
\item{pvalues}{pvalues}

\item{alpha}{alpha level, default 0.05}

\item{lambda}{lambda}

\item{family}{family}
}
\value{
Returns pvalues
}
\description{
cvQuantile
}
\author{
Angela Andreella
}
