% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaOptQuantile.R
\name{lambdaOptQuantile}
\alias{lambdaOptQuantile}
\title{lambdaOptQuantile}
\usage{
lambdaOptQuantile(B, pvalues, m, alpha)
}
\arguments{
\item{B}{number of permutation, default 1000}

\item{pvalues}{pvalues}

\item{alpha}{alpha level, default 0.05}

\item{family}{family}
}
\value{
Returns pvalues
}
\description{
lambdaOptQuantile
}
\author{
Angela Andreella
}
