% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxTQuantile.R
\name{maxTQuantile}
\alias{maxTQuantile}
\title{maxT quantile approach}
\usage{
maxTQuantile(X, B, alpha, label)
}
\arguments{
\item{X}{data where rows represents the variables and columns the observations, the columns's name define the categories}

\item{B}{number of permutations to perform, default is 1000.}

\item{alpha}{alpha level}

\item{label}{labels of the observations, if NULL the columns's name are considered, default NULL}
}
\value{
Returns the tests rejected by maxT Westfall and Young with respective indeces.
}
\description{
Performs maxT quantile approach
}
\author{
Angela Andreella
}
