\name{qpARI-package}
\alias{qpARI-package}
\alias{qpARI}
\docType{package}
\title{
\packageTitle{qpARI}
}
\description{
\packageDescription{qpARI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{qpARI}
\packageIndices{qpARI}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{qpARI}

Maintainer: \packageMaintainer{qpARI}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
